requirements.txt
alembic
boto3
celery
email-validator
Flask
Flask-Caching
Flask-Limiter
Flask-Login
Flask-Migrate
Flask-Talisman
Flask-Babel
Flask-WTF
flask-sqlalchemy
filetype
google-generativeai
google-genai
gunicorn
itsdangerous
passlib[bcrypt]
pydantic
Pillow
python-dotenv
requests
SQLAlchemy
watchdog
waitress
twilio
razorpay
